+++ DFA/project.c
@@ -18,6 +18,7 @@
  * USA.
  */
 
+#include <stdint.h>
 #include "dfa.h"
 #include "../BDD/hash.h"
 #include "../Mem/mem.h"
@@ -48,7 +49,7 @@ void init_ssets(int sz)
    next_sset = 0;
 }
 
-int make_sset(int sz, int *elem,  unsigned sq, int d1, int d2)
+int make_sset(int sz, int *elem, unsigned sq, int d1, int d2)
 {
   if (next_sset == n_ssets) {
     struct set *new_ssets = mem_alloc((sizeof *ssets) * n_ssets * 2);
@@ -64,7 +65,7 @@ int make_sset(int sz, int *elem,  unsigned sq, int d1, int d2)
   ssets[next_sset].decomp1 = d1;
   ssets[next_sset].decomp2 = d2;
   ssets[next_sset].permanent = -1;
-  insert_in_hash_tab(htbl_set, (long)elem, 0, (void *)(next_sset+1));  
+  insert_in_hash_tab(htbl_set, (long)elem, 0, (void *)(uintptr_t)(next_sset+1));  
   /* htbl maps to ++id, since 0 = not_found */ 
 
   return(next_sset++);
@@ -120,7 +121,7 @@ unsigned proj_term1(unsigned state1, unsigned  state2)
   }
    
   /* res = 0 or id+1 */
-  if ( (res = (int) lookup_in_hash_tab(htbl_set, (long) s, 0)) ) {
+  if ( (res = (int)(uintptr_t) lookup_in_hash_tab(htbl_set, (long) s, 0)) ) {
     mem_free(s); /* it was already there */  
     return (--res);
   }
@@ -166,7 +167,7 @@ bdd_ptr proj_term2(unsigned set_index1,  unsigned set_index2)
   *e3 = -1;   /* Terminate the new set */
   
   /* res = 0 or id+1 */
-  if ( (res = (int) lookup_in_hash_tab(htbl_set, (long) s, 0)) ) {
+  if ( (res = (int)(uintptr_t) lookup_in_hash_tab(htbl_set, (long) s, 0)) ) {
     mem_free(s); /* it was already there */
     return (--res);
   }
